% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTIS_available_country.R
\name{CTIS_available_country}
\alias{CTIS_available_country}
\title{Check availability of CTIS data for a country}
\usage{
CTIS_available_country(country)
}
\arguments{
\item{country}{Country, e.g., "Germany"}
}
\value{
A data frame object that indicates the dates for which data from
the Global COVID-19 Trends and Impact Survey data is available for the
respective country
}
\description{
Provides a list of dates for which data from the Global
COVID-19 Trends and Impact Survey for the specified country is available
}
\examples{
Germany_available_dates <- CTIS_available_country("Germany")
head(Germany_available_dates)
}
