% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTIS_available_region.R
\name{CTIS_available_region}
\alias{CTIS_available_region}
\title{Check availability of CTIS data for a region in a country}
\usage{
CTIS_available_region(country, region)
}
\arguments{
\item{country}{Country, e.g., "Germany"}

\item{region}{Region, e.g., "Bayern" (usually in native language)}
}
\value{
A data frame object that indicates the dates for which data from the
Global COVID-19 Trends and Impact Survey data is available for the respective
country region
}
\description{
Provides a list of dates for which Global COVID-19 Trends and
Impact Survey data is available for a region in a country
}
\examples{
Bavaria_available_dates <- CTIS_available_region("Germany", "Bayern")
head(Bavaria_available_dates)
}
