% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTIS_microdata.R
\name{CTIS_microdata}
\alias{CTIS_microdata}
\title{CTIS_microdata}
\usage{
CTIS_microdata(
  username,
  password,
  date,
  surveyversion = "v1.7",
  type = "parta"
)
}
\arguments{
\item{username}{Usernme for the microdata API}

\item{password}{Password for the microdata API}

\item{date}{Date in the format "YYYY-MM-DD"}

\item{surveyversion}{Current version of the Global COVID-19 Trends and Impact
Survey: July 2021: v1.7}

\item{type}{"full" or "partA" version}
}
\value{
A data frame object that includes the microdata data for the date
specified from the Global COVID-19 Trends and Impact Survey
}
\description{
Provides microdata for the date specified from the Global
COVID-19 Trends and Impact Survey
}
