% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTIS_open_data_country.R
\name{CTIS_open_data_country}
\alias{CTIS_open_data_country}
\title{CTIS_open_data_country}
\usage{
CTIS_open_data_country(indicator, type, country, daterange)
}
\arguments{
\item{indicator}{Indicator variable, e.g., "vaccine_acpt",
"covid_vaccine", etc.}

\item{type}{Data type, "daily" or "smoothed"}

\item{country}{Country, e.g., "Germany"}

\item{daterange}{A particular date, e.g., "20210101"or a date range
"20210501-20210516"}
}
\value{
A data frame object that includes the indicator data for the country
and date specified from the Global COVID-19 Trends and Impact Survey
}
\description{
Provides data for the indicator, country, date and type
specified from the Global COVID-19 Trends and Impact Survey
}
\examples{
Germany_vaccine_acceptance_20210501_20210516 <-
CTIS_open_data_country("vaccine_acpt",
"daily",
"Germany",
"20210501-20210516")
head(Germany_vaccine_acceptance_20210501_20210516)
}
