% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTIS_open_data_region.R
\name{CTIS_open_data_region}
\alias{CTIS_open_data_region}
\title{CTIS_open_data_region}
\usage{
CTIS_open_data_region(indicator, type, country, region, daterange)
}
\arguments{
\item{indicator}{Indicator variable, e.g., "vaccine_acpt", "covid_vaccine",
etc. Please refer to
https://gisumd.github.io/COVID-19-API-Documentation/docs/indicators/indicators.html}

\item{type}{Data type, "daily" or "smoothed"}

\item{country}{Country, e.g., "Germany"}

\item{region}{Region, e.g., "Bayern" (usually in native language)}

\item{daterange}{A particular date in format YYYYMMDD,e.g., "20210101" or
a date range, e.g., "20210501-20210516"}
}
\value{
A data frame object that includes the indicator data for the country,
region and date specified from the Global COVID-19 Trends and Impact Survey
}
\description{
Provides data for the indicator, country, region date and type
specified from the Global COVID-19 Trends and Impact Survey
}
\examples{
Bavaria_vaccine_acceptance_20210501_20210516 <-
CTIS_open_data_region("vaccine_acpt",
"daily",
"Germany",
"Bayern",
"20210501-20210516")

head(Bavaria_vaccine_acceptance_20210501_20210516)
}
