% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTIS_survey_counts.R
\name{CTIS_survey_counts}
\alias{CTIS_survey_counts}
\title{CTIS_survey_counts}
\usage{
CTIS_survey_counts(date)
}
\arguments{
\item{date}{A particular date in format YYYYMMDD, e.g., "20210101",}
}
\value{
A data frame object that includes the number of responses for a
specific date from the Global COVID-19 Trends and Impact Survey
}
\description{
Provides data for the number of full and PartA responses for a
specific date from the Global COVID-19 Trends and Impact Survey
}
\examples{
CTIS_survey_counts("20210129")
}
