% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTIS_microdata.R
\name{CTIS_microdata}
\alias{CTIS_microdata}
\title{CTIS_microdata}
\usage{
CTIS_microdata(
  username,
  password,
  date,
  surveyversion = "v1.7",
  type = "parta"
)
}
\arguments{
\item{username}{Username for the microdata API, please check
https://dataforgood.fb.com/docs/covid-19-symptom-survey-request-for-data-access/
if you do not already have an account}

\item{password}{Password for the microdata API, please check
https://dataforgood.fb.com/docs/covid-19-symptom-survey-request-for-data-access/
if you do not already have an account}

\item{date}{Date in the format "YYYY-MM-DD"}

\item{surveyversion}{Current version of the Global COVID-19 Trends and Impact
Survey. Current version during July 2021: v1.7}

\item{type}{"full" or "parta" version}
}
\value{
A data frame object that includes the microdata data for the date
specified from the Global COVID-19 Trends and Impact Survey
}
\description{
Downloads, saves as a .csv in the working directory and then
loads the microdata for thedate specified from the Global COVID-19 Trends and
Impact Survey
}
